Improved _run_bot()'s info and changed _reply()
_run_bot() previously would say "Replied to: [submission_information]"
even if the bot had already responded to that submission. Now the bot
will print the same message _debug() prints when the bot comes across a
submission already in repliedto.txt. This makes _run_bot() clearer to
understand.

_reply() will no longer check if a submission was already replied to.
That task has been delegated to the _debug() and _run_bot() functions,
which now call the particular helper function that _reply() used to call.