Changed the TRIGGER_PATTERN
The bot will now try to account for submission authors who are transfers
to UCI for the CS major, but aren't in the ICS school anymore. It will
try to not respond to submissions of this sort, but there are some it
will still respond to at the time of this writing.