Improved TRIGGER_PATTERN
The bot will now be able to deal with phrases such as "transferring to
CS." Previously I had a typo in the "transfer" part of the regular
expression (I forgot an "r" for the "(ing)?" group), so the bot would
detect "transfering" (a mispelling) rather than "transferring" (the
correct spelling). I also added a "to" alternative in the "transferring" section as well.

In short, the bot's "transferring" section is more robust than before.