Improved performance of the help message
Previously the bot would always initialize itself using PRAW so that it
could post on Reddit and see subreddit submissions, even if the bot was
run with no command-line arguments (at the time of this commit, this
means the bot should only generate a help message, nothing more). Of
course, if all the user wanted was to see the help message for the bot,
this would be quite inefficient.

The bot will now only do all its PRAW stuff when it is running with
either the -r or -d flags.